MODRS-5 Fixed tests
