--- src/main/java/org/folio/rs/domain/entity/DomainEventType.java
@@ -1,5 +1,0 @@
-package org.folio.rs.domain.entity;
-
-public enum DomainEventType {
-  UPDATE
-}
